backend/requirements.txt
# requirements.txt
Django==4.2
djangorestframework==3.14.0
channels==4.0.0
celery==5.3.0
redis==4.5.5
psycopg2-binary==2.9.6
Pillow==10.0.0